backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.40.0
pandas==2.2.3
numpy==2.2.2
python-dotenv==1.0.1
requests==2.32.3

